LayoutTests/http/wpt/navigation-api/traverse-subframe-remote-sibling.sub.html

<!doctype html><!-- webkit-test-runner [ SiteIsolationEnabled=true ] -->
<meta charset="utf-8">
<title>Navigation API: subframe traversal must not abandon a local sibling frame when a cross-origin RemoteFrame sibling precedes it (Site Isolation)</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<!--
  Issue C repro. Frame tree (child order matters):
    main (primary origin, stays same-document)
      |-- iframeRemote  ({{hosts[alt][]}}, cross-site) -> RemoteFrame under SI
      |-- iframeLocal   (primary origin, same-site)    -> LocalFrame under SI

  iframeRemote is appended FIRST so it is the first child in the main frame's
  tree. iframeLocal is navigated cross-document (?p=0 -> 1 -> 2) to build extra
  back/forward entries, then a joint traversal moves it back to ?p=1.

  On today's code (origin/main), HistoryController::recursiveGatherFramesToNavigate
  (Source/WebCore/loader/HistoryController.cpp:501-503) does, for the first child:
        RefPtr subframe = dynamicDowncast<LocalFrame>(descendantByFrameID(id));
        if (!subframe)
            return;   // BUG: bare return abandons the ENTIRE sibling loop
  The RemoteFrame downcast is null and the bare return abandons iframeLocal too,
  so iframeLocal never gets loadItem() and does not traverse. Sibling functions
  recursiveSetProvisionalItem (977) and recursiveGoToItem (999) use skip semantics
  (if (LocalFrame) {...}); only the gather loop returns. Fix: return -> continue.

  Expected today: FAIL/timeout -- iframeLocal stays at index 2 (?p=2).
  With the fix: PASS -- iframeLocal traverses back to index 1 (?p=1).
-->
<script>
const REMOTE = "http://{{hosts[alt][]}}:{{ports[http][0]}}/WebKit/navigation-api/resources/traverse-subframe-child.html?p=remote";
const LOCAL = "resources/traverse-subframe-child.html";

function makeIframe(name, src) {
  return new Promise(resolve => {
    const f = document.createElement("iframe");
    f.name = name;
    f.src = src;
    f.onload = () => resolve(f);
    document.body.appendChild(f);
  });
}

function navigateLocal(iframe, url) {
  return new Promise(resolve => {
    iframe.onload = () => resolve();
    iframe.contentWindow.location.href = url;
  });
}

// Poll until iframeLocal's Navigation currentEntry reaches the expected index.
// Used instead of awaiting traverseTo().committed so the test measures the
// observable navigation directly (see comment at the traversal site).
function waitForLocalIndex(expected, timeoutMs) {
  return new Promise((resolve, reject) => {
    const start = Date.now();
    const tick = () => {
      const nav = iframeLocal.contentWindow.navigation;
      if (nav && nav.currentEntry && nav.currentEntry.index === expected) {
        resolve();
        return;
      }
      if (Date.now() - start >= timeoutMs) {
        reject(new Error(`timed out waiting for iframeLocal index ${expected}; still at ${nav && nav.currentEntry ? nav.currentEntry.index : "?"}`));
        return;
      }
      setTimeout(tick, 25);
    };
    tick();
  });
}

let iframeRemote, iframeLocal;

promise_test(async t => {
  // Remote first => first child in the main frame's tree.
  iframeRemote = await makeIframe("iframeRemote", REMOTE);
  iframeLocal = await makeIframe("iframeLocal", LOCAL + "?p=0");

  // Let the initial loads settle so subsequent navigations are push, not replace.
  await new Promise(r => t.step_timeout(r, 0));

  // Build several cross-document entries in iframeLocal. Distinct URLs push.
  // We do not hardcode the resulting count (the initial-load entry may or may
  // not be replaced by the first navigation); we adapt to whatever the local
  // frame's session history ends up being.
  await navigateLocal(iframeLocal, LOCAL + "?p=1");
  await navigateLocal(iframeLocal, LOCAL + "?p=2");
  await navigateLocal(iframeLocal, LOCAL + "?p=3");

  const localNav = iframeLocal.contentWindow.navigation;
  const startIndex = localNav.currentEntry.index;
  const startSearch = iframeLocal.contentWindow.location.search;
  assert_greater_than_equal(startIndex, 1,
    `iframeLocal must have at least one back entry (index=${startIndex}, len=${localNav.entries().length})`);

  // The immediately-previous entry is our traversal target.
  const target = localNav.entries()[startIndex - 1];
  const targetKey = target.key;
  const targetUrl = new URL(target.url);

  // Joint traversal back one entry. Dispatches through goToItemForNavigationAPI
  // and recursiveGatherFramesToNavigate rooted at the MAIN frame, which hits the
  // RemoteFrame sibling first. Under the bug the gather loop bare-returns on the
  // RemoteFrame and iframeLocal is never navigated at all: its index stays put.
  //
  // We assert on the observable navigation (index + URL) rather than on the
  // traverseTo() promise: whether the joint traverse's committed/finished promise
  // settles for a subframe target under Site Isolation is a separate concern
  // (async cross-frame promise settlement; see research/06) and is not what this
  // abandonment fix addresses.
  iframeLocal.contentWindow.navigation.traverseTo(targetKey);

  await waitForLocalIndex(startIndex - 1, 4000);

  assert_equals(iframeLocal.contentWindow.navigation.currentEntry.index, startIndex - 1,
    `iframeLocal traversed back one entry (${startIndex} -> ${startIndex - 1}) despite the preceding RemoteFrame sibling`);
  assert_equals(iframeLocal.contentWindow.location.search, targetUrl.search,
    `iframeLocal document moved to the previous entry's URL (${startSearch} -> ${targetUrl.search})`);
}, "Subframe traversal re-navigates a same-origin LocalFrame even when a cross-origin RemoteFrame sibling precedes it (Site Isolation)");
</script>
</body>
